# source/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/dataSet/add": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataSet"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DataSet"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dataSet/delete": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataSet"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dataSet/get-all": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataSet"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DataSet_Search"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/model.DataSet_View"
                                    }
                                },
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dataSet/get-one": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataSet"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/model.DataSet_View"
                                },
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/dataSet/update": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dataSet"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DataSet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/public/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User_Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/public/auth/refresh": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/public/auth/refresh2": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/public/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/public/user/reset-password": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User_ResetPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/reset-password": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User_ResetPassword_Admin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.DataSet": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "column": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "createdAt": {
                    "type": "integer"
                },
                "data_description": {
                    "type": "string"
                },
                "data_name": {
                    "type": "string"
                },
                "data_source": {
                    "type": "string"
                },
                "data_table": {
                    "type": "boolean"
                },
                "is_required_id": {
                    "type": "boolean"
                },
                "is_required_province_code": {
                    "type": "boolean"
                },
                "is_required_province_name": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "upload_file": {
                    "type": "boolean"
                },
                "value_row": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "model.DataSet_Search": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "order": {
                    "type": "string",
                    "example": "DESC"
                },
                "orderBy": {
                    "type": "string",
                    "example": "createdAt"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "range": {
                    "$ref": "#/definitions/model.Range"
                },
                "search": {
                    "description": "? Regex",
                    "type": "string"
                },
                "searchBy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size": {
                    "type": "integer",
                    "example": 11
                }
            }
        },
        "model.DataSet_View": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "column": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "createdAt": {
                    "type": "integer"
                },
                "data_description": {
                    "type": "string"
                },
                "data_name": {
                    "type": "string"
                },
                "data_source": {
                    "type": "string"
                },
                "data_table": {
                    "type": "boolean"
                },
                "is_required_id": {
                    "type": "boolean"
                },
                "is_required_province_code": {
                    "type": "boolean"
                },
                "is_required_province_name": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "upload_file": {
                    "type": "boolean"
                },
                "value_row": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "model.MetadataResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/model.PaginationResponse"
                },
                "status": {
                    "type": "boolean"
                },
                "timeExecution": {
                    "type": "string"
                }
            }
        },
        "model.PaginationResponse": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                },
                "totalElements": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "model.Range": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer",
                    "example": 1646792565000
                },
                "field": {
                    "type": "string",
                    "example": "updatedAt"
                },
                "start": {
                    "type": "integer",
                    "example": 1646792565000
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "fullname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roleId": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.User_Login": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "model.User_ResetPassword": {
            "type": "object",
            "properties": {
                "confirmpassword": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                },
                "oldPassword": {
                    "type": "string"
                }
            }
        },
        "model.User_ResetPassword_Admin": {
            "type": "object",
            "properties": {
                "NewPassword": {
                    "type": "string"
                },
                "confirmpassword": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "description": "E.g. Bearer Your.Token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}